Modified RTCBuilder GUI.#411
