+++ jp.go.aist.rtm.rtcbuilder.nl1/src/jp/go/aist/rtm/rtcbuilder/nl/messages_ja.properties
@@ -345,24 +345,11 @@ IMessageConstants.DOCUMENT_HINT_CREATOR_DESC_P1=name <mail address> \u306e\u66f8
 IMessageConstants.DOCUMENT_HINT_CREATOR_DESC_P2=\u540d\u524d[name]\u306f\u30ed\u30fc\u30de\u5b57\u8868\u8a18\u3067\u5165\u529b\u3057\u307e\u3059\u3002
 IMessageConstants.DOCUMENT_HINT_CREATOR_DESC_P3=\u30e1\u30fc\u30eb\u30a2\u30c9\u30ec\u30b9\u306f<>\u8a18\u53f7\u3067\u62ec\u308b\u5fc5\u8981\u304c\u3042\u308a\u307e\u3059\u3002
 
-IMessageConstants.LANGUAGE_SECTION=\u8a00\u8a9e\u30fb\u74b0\u5883
 IMessageConstants.LANGUAGE_LANG_TITLE=\u8a00\u8a9e
-IMessageConstants.LANGUAGE_ENV_TITLE=\u74b0\u5883
 IMessageConstants.LANGUAGE_LANG_EXPL=\u3053\u306e\u30bb\u30af\u30b7\u30e7\u30f3\u3067\u306f\u4f7f\u7528\u3059\u308b\u8a00\u8a9e\u3092\u6307\u5b9a\u3057\u307e\u3059
-IMessageConstants.LANGUAGE_ENV_EXPL=\u3053\u306e\u30bb\u30af\u30b7\u30e7\u30f3\u3067\u306f\u4f9d\u5b58\u3059\u308b\u30e9\u30a4\u30d6\u30e9\u30ea\u3084\u4f7f\u7528\u3059\u308bOS\u306a\u3069\u3092\u6307\u5b9a\u3057\u307e\u3059
 
 IMessageConstants.LANGUAGE_HINT_LANG_TITLE=\u8a00\u8a9e\uff1a
 IMessageConstants.LANGUAGE_HINT_LANG_DESC=RT\u30b3\u30f3\u30dd\u30fc\u30cd\u30f3\u30c8\u3092\u4f5c\u6210\u3059\u308b\u8a00\u8a9e\u3092\u9078\u629e\u3057\u307e\u3059\u3002\u30ea\u30b9\u30c8\u4e2d\u306e\u8a00\u8a9e\u304b\u3089\u9078\u629e\u53ef\u80fd\u3067\u3059\u3002
-IMessageConstants.LANGUAGE_HINT_ENV_TITLE=\u74b0\u5883\uff1a
-IMessageConstants.LANGUAGE_HINT_ENV_DESC_P1=\u8a00\u8a9e\u3054\u3068\u306e\u30e9\u30a4\u30d6\u30e9\u30ea\u306e\u4f9d\u5b58\u95a2\u4fc2\u3084\u3001\u4f7f\u7528\u3059\u308bOS\u306a\u3069\u306e\u74b0\u5883\u3092\u9078\u629e\u3057\u307e\u3059\u3002
-IMessageConstants.LANGUAGE_HINT_ENV_DESC_P2=
-IMessageConstants.LANGUAGE_HINT_ENV_DESC_P3=\u8a73\u7d30\u60c5\u5831\u3067\u8a2d\u5b9a\u3057\u305f\u5185\u5bb9\uff08OS\u60c5\u5831\u3001\u30e9\u30a4\u30d6\u30e9\u30ea\u60c5\u5831\u306a\u3069\uff09\u306f\u3001\u30d7\u30ed\u30d5\u30a1\u30a4\u30eb\u5185\u306b\u306e\u307f\u4fdd\u5b58\u3055\u308c\u307e\u3059\u3002
-IMessageConstants.LANGUAGE_HINT_ENV_DESC_P4=
-
-IMessageConstants.LANGUAGE_LBL_DETAIL=\u8a73\u7d30\u60c5\u5831
-IMessageConstants.LANGUAGE_LBL_OSETC=\u305d\u306e\u4ed6OS\u60c5\u5831
-IMessageConstants.LANGUAGE_LBL_CPUETC=\u305d\u306e\u4ed6CPU\u60c5\u5831
-IMessageConstants.LANGUAGE_LBL_LIBRARY=\u30e9\u30a4\u30d6\u30e9\u30ea\u60c5\u5831
 
 IMessageConstants.LANGUAGE_SELECTION_CAUTION=\u751f\u6210\u5bfe\u8c61\u8a00\u8a9e\u3092\u9078\u629e\u3057\u3066\u304f\u3060\u3055\u3044\u3002
 IMessageConstants.LANGUAGE_CHOREONOID_CAUTION=Choreonoid\u578bRTC\u306fC++\u306e\u307f\u9078\u629e\u53ef\u80fd\u3067\u3059\u3002
@@ -504,7 +491,7 @@ IPreferenceMessageConstants.CONFIG_BTN_DELETE=Delete
 
 IPreferenceMessageConstants.DOCUMENT_LBL_AUTHOR=\u4f5c\u6210\u8005,\u9023\u7d61\u5148
 IPreferenceMessageConstants.DOCUMENT_LBL_LICENSE_P1=\u30e9\u30a4\u30bb\u30f3\u30b9,\u4f7f\u7528\u6761\u4ef6
-IPreferenceMessageConstants.DOCUMENT_LBL_LICENSE_P2=\u53c2\u8003\u6587\u732e
+IPreferenceMessageConstants.DOCUMENT_LBL_LICENSE_P2=
 
 IPreferenceMessageConstants.PORT_TITLE_DATA_PORT=\u30c7\u30fc\u30bf\u30dd\u30fc\u30c8
 IPreferenceMessageConstants.PORT_TITLE_SERVICE_PORT=\u30b5\u30fc\u30d3\u30b9\u30dd\u30fc\u30c8
