+++ jp.go.aist.rtm.rtcbuilder/src/jp/go/aist/rtm/rtcbuilder/nl/messages.properties
@@ -345,24 +345,11 @@ IMessageConstants.DOCUMENT_HINT_CREATOR_DESC_P1=The input format is name <mail a
 IMessageConstants.DOCUMENT_HINT_CREATOR_DESC_P2=
 IMessageConstants.DOCUMENT_HINT_CREATOR_DESC_P3=
 
-IMessageConstants.LANGUAGE_SECTION=Language and Environment
 IMessageConstants.LANGUAGE_LANG_TITLE=Language
-IMessageConstants.LANGUAGE_ENV_TITLE=Environment
 IMessageConstants.LANGUAGE_LANG_EXPL=This section defines a language that is used.
-IMessageConstants.LANGUAGE_ENV_EXPL=This section defines depending libraries and Oses etc that are used.
 
 IMessageConstants.LANGUAGE_HINT_LANG_TITLE=Language :
 IMessageConstants.LANGUAGE_HINT_LANG_DESC=Selects a programming language used to develop RT-Component.
-IMessageConstants.LANGUAGE_HINT_ENV_TITLE=Environment :
-IMessageConstants.LANGUAGE_HINT_ENV_DESC_P1=Selects the dependency in library of each language and the environment
-IMessageConstants.LANGUAGE_HINT_ENV_DESC_P2=such as OS that is used.
-IMessageConstants.LANGUAGE_HINT_ENV_DESC_P3=The set contents (Oses and Libraries etc) in the detail information
-IMessageConstants.LANGUAGE_HINT_ENV_DESC_P4=is saved only within Profile.
-
-IMessageConstants.LANGUAGE_LBL_DETAIL=Detail information
-IMessageConstants.LANGUAGE_LBL_OSETC=Other Oses
-IMessageConstants.LANGUAGE_LBL_CPUETC=Other CPUs
-IMessageConstants.LANGUAGE_LBL_LIBRARY=Libraries
 
 IMessageConstants.LANGUAGE_SELECTION_CAUTION=Please select a language for generation.
 IMessageConstants.LANGUAGE_CHOREONOID_CAUTION=Choreonoid type RTC can only select C++.
